+++ app/models/contact.rb
@@ -1,0 +1,4 @@
+class Contact < ApplicationRecord
+  validates :email, uniqueness: true
+  validates :email, :name, presence: true
+end