+++ app/models/track.rb
@@ -1,0 +1,3 @@
+class Track < ApplicationRecord
+  validates :guid, :visited_page, :visited_datetime, presence: true
+end