+++ spec/factories/tracks.rb
@@ -1,0 +1,12 @@
+FactoryBot.define do
+  factory :track do
+    guid { SecureRandom.uuid }
+    visited_page { select_page }
+    visited_datetime { Faker::Time.between(2.days.ago,Date.today, :all) }
+  end
+
+  def select_page
+    page_list = ['Home', 'Blog', 'News', 'About', 'Contact']
+    page_list[rand(page_list.length)]
+  end
+end