+++ spec/rails_helper.rb
@@ -27,7 +27,7 @@ require 'rspec/rails'
 # directory. Alternatively, in the individual `*_spec.rb` files, manually
 # require only the support files necessary.
 #
-# Dir[Rails.root.join('spec', 'support', '**', '*.rb')].each { |f| require f }
+Dir[Rails.root.join('spec', 'support', '**', '*.rb')].each { |f| require f }
 
 # Checks for pending migrations and applies them before tests are run.
 # If you are not using ActiveRecord, you can remove these lines.
@@ -78,6 +78,7 @@ RSpec.configure do |config|
 
   # Personal Configs
   config.include FactoryBot::Syntax::Methods
+  config.include RequestHelper, type: :request
 
   config.before(:suite) do
     DatabaseCleaner.clean_with(:truncation)