+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-# ContactList 0.0.1 (Apr 04, 2019)
+# ContactList 0.0.1 (Apr 06, 2019)
 
-* TDD: Added tests to create models
+* TDD: Added tests to create models:
+* Contact and Track
 
   *Marcelo Toledo*